LayoutTests/editing/pasteboard/paste-text-with-style-4-expected.txt

This tests copying and pasting text does not remove non-editing style in the destination.
To manually test, copy "hello world" below and paste it in the box below. The blue border should stay after the paste.

Before copy:
| <b>
|   style="border: solid 5px blue;padding: 5px;"
|   <br>

After paste:
| <b>
|   style="border: solid 5px blue;padding: 5px;"
|   <span>
|     class="Apple-style-span"
|     style="font-weight: normal; "
|     <i>
|       "hello<#selection-caret>"
|   <br>
